fix(client uuid): client uuid is 8 bytes instead of 4
* Fix NullPointerException where cache was accessed while being initialized.

* Fix URISyntaxException illegal character in module path.

* Ignore libraries folder.

* Changed UUID from int to long

+++ src/main/java/org/maxgamer/rs/logon/logon/GameDecoder.java
@@ -29,7 +29,7 @@ public class GameDecoder extends OpcodeDecoder<LSIncomingPacket> implements Hand
         String name = in.readPJStr1();
         String pass = in.readPJStr1();
         String ip = in.readPJStr1();
-        int clientUUID = in.readInt();
+        long clientUUID = in.readLong();
 
         LSOutgoingPacket out = new LSOutgoingPacket(3);
         AuthResult result = AuthResult.SUCCESS;
